refine code for readability
