+++ Cargo.toml
@@ -221,7 +221,7 @@ procinfo = { git = "https://github.com/tikv/procinfo-rs", rev = "6599eb9dca74229
 # kvproto at the same time.
 # After the PR to kvproto is merged, remember to comment this out and run `cargo update -p kvproto`.
 [patch.'https://github.com/pingcap/kvproto']
-kvproto = { git = "https://github.com/JmPotato/kvproto", branch="move_flashback_start_ts" }
+kvproto = { git = "https://github.com/JmPotato/kvproto", branch = "move_flashback_start_ts" }
 
 [workspace]
 # See https://github.com/rust-lang/rfcs/blob/master/text/2957-cargo-features2.md
